backend/requirements.txt
Flask==2.2.5
Flask-Cors==3.0.10
Flask-SQLAlchemy==3.0.3
python-dotenv==1.0.0
PyMySQL==1.0.3
bcrypt==4.0.1
paho-mqtt==1.6.1
twilio==8.0.0
stripe==5.0.0
PyJWT==2.8.0
requests==2.31.0
